EditorBuffer: decompile setupParamDefineTable
Other Changes:
- Change return type of System::getUserParamNum from u64 to u32
- Added getter for mIsInitialized in ParamDefineTable

+++ include/xlink2/xlink2SystemELink.h
@@ -52,7 +52,7 @@ public:
                                         sead::SafeString const&, Event*, AssetExecutorELink*,
                                         bool) const;
 
-    u64 getUserParamNum() const override;
+    u32 getUserParamNum() const override;
     ILockProxy* getModuleLockObj() const override;
 
 private:
