+++ src/xlink2SystemELink.cpp
@@ -5,7 +5,7 @@ u64 SystemELink::getResourceVersion() const {
     return 0x1e;
 }
 
-u64 SystemELink::getUserParamNum() const {
+u32 SystemELink::getUserParamNum() const {
     return 0;
 }
